+++ servant-client/CHANGELOG.md
@@ -3,6 +3,7 @@ HEAD
 
 * Support for the `HttpVersion`, `IsSecure`, `RemoteHost` and `Vault` combinators
 * Added support for `path` on `BaseUrl`.
+* `client` now takes an explicit `Manager` argument.
 
 0.4.1
 -----